+++ src/test/java/nextstep/member/acceptance/AuthAcceptanceTest.java
@@ -10,6 +10,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import static nextstep.member.acceptance.AuthSteps.*;
+import static nextstep.member.acceptance.MemberSteps.내_정보_요청;
 import static org.assertj.core.api.Assertions.assertThat;
 
 class AuthAcceptanceTest extends AcceptanceTest {
