+++ src/test/java/nextstep/member/acceptance/MemberAcceptanceTest.java
@@ -7,7 +7,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 
-import static nextstep.member.acceptance.AuthSteps.내_정보_요청;
 import static nextstep.member.acceptance.AuthSteps.토큰_생성;
 import static nextstep.member.acceptance.MemberSteps.*;
 import static org.assertj.core.api.Assertions.assertThat;
